Added information about using github action (#1226)
By using this Github action it is possible to keep the derived
repository up-to-date with its template.

I am currently using this to keep my website up to date with the
template repository. An example of the PR this action creates can be
seen [here](https://github.com/george-gca/george-gca.github.io/pull/13).

Signed-off-by: George Araújo <george.gcac@gmail.com>